Commit Changes: - Split in 2 the initial login sequence     1 - Creates the User instance in my db     2 - Redirects to the dashboard depending the RoleRedirect - Added a Spinner component for loading screens
This will help identify errors easily in the future
