+++ frontend/src/pages/RoleRedirect.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { useEffect } from "react";
+import { Spinner } from "@heroui/react";
 import { useAuth0 } from "@auth0/auth0-react";
 import { useNavigate } from "react-router-dom";
 
@@ -9,25 +10,6 @@ export default function RoleRedirect() {
   const navigate = useNavigate();
 
   useEffect(() => {
-    async function createUser(params) {
-      let userInfo = {
-        name: user.given_name,
-        lastname: user.family_name,
-        email: user.email,
-        birthDate: new Date(),
-      };
-
-      await fetch(import.meta.env.VITE_BASE_URL + "/user/create-new-user", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(userInfo),
-      });
-    }
-
-    createUser();
-
     if (isAuthenticated) {
       const roleArr = user["https://hospitall.com/roles"];
 
@@ -41,5 +23,9 @@ export default function RoleRedirect() {
     }
   }, []);
 
-  return <div>RoleRedirect</div>;
+  return (
+    <div className="w-screen h-screen flex items-center justify-center">
+      <Spinner color="primary" label="Redirecting..." className="text-white" />
+    </div>
+  );
 }
